serialization: add shema processors
