+++ invenio_rdm_records/contrib/meeting/__init__.py
@@ -3,6 +3,7 @@
 #
 # Invenio-RDM-Records is free software; you can redistribute it and/or modify
 # it under the terms of the MIT License; see LICENSE file for more details.
+
 """Meeting contrib module."""
 
 from .custom_fields import (
