+++ invenio_rdm_records/resources/serializers/datacite/__init__.py
@@ -10,12 +10,8 @@ from datacite import schema43
 from flask_resources import BaseListSchema, MarshmallowSerializer
 from flask_resources.serializers import JSONSerializer, SimpleSerializer
 
-from invenio_rdm_records.contrib.journal.custom_fields import JournalDataciteDumper
-
+from ....contrib.journal.processors import JournalDataciteDumper
 from .schema import DataCite43Schema
-
-# Order matters
-datacite_dump_dumpers = [JournalDataciteDumper()]
 
 
 class DataCite43JSONSerializer(MarshmallowSerializer):
@@ -27,7 +23,7 @@ class DataCite43JSONSerializer(MarshmallowSerializer):
             format_serializer_cls=JSONSerializer,
             object_schema_cls=DataCite43Schema,
             list_schema_cls=BaseListSchema,
-            schema_kwargs={"dumpers": datacite_dump_dumpers},
+            schema_kwargs={"dumpers": [JournalDataciteDumper()]},  # Order matters
             **options
         )
 
@@ -42,6 +38,6 @@ class DataCite43XMLSerializer(MarshmallowSerializer):
             format_serializer_cls=SimpleSerializer,
             object_schema_cls=DataCite43Schema,
             list_schema_cls=BaseListSchema,
-            schema_kwargs={"dumpers": datacite_dump_dumpers},
+            schema_kwargs={"dumpers": [JournalDataciteDumper()]},  # Order matters
             encoder=encoder,
         )
