+++ invenio_rdm_records/resources/serializers/dublincore/schema.py
@@ -9,16 +9,17 @@
 
 import bleach
 import idutils
+from flask_resources.serializers import BaseSerializerSchema
 from invenio_access.permissions import system_identity
 from invenio_vocabularies.proxies import current_service as vocabulary_service
 from marshmallow import fields, missing
 
-from ..schemas import CommonFieldsMixin, DumperMixin
+from ..schemas import CommonFieldsMixin
 from ..ui.schema import current_default_locale
 from ..utils import get_vocabulary_props
 
 
-class DublinCoreSchema(CommonFieldsMixin, DumperMixin):
+class DublinCoreSchema(BaseSerializerSchema, CommonFieldsMixin):
     """Schema for Dublin Core in JSON.
 
     The identifier handling behavior is determined by the schema's ``context``,
