+++ invenio_rdm_records/resources/serializers/schemas.py
@@ -6,12 +6,13 @@
 # it under the terms of the MIT License; see LICENSE file for more details.
 
 """Base parsing functions for the various serializers."""
-from marshmallow import Schema, missing, post_dump, post_load, pre_dump, pre_load
 
-from invenio_rdm_records.contrib.journal.custom_fields import FieldDumper, FieldLoader
+from marshmallow import missing
 
 
-class CommonFieldsMixin(object):
+class CommonFieldsMixin:
+    """Common fields serialization."""
+
     def get_doi(self, obj):
         """Get DOI."""
         if "doi" in obj["pids"]:
@@ -79,100 +80,3 @@ class CommonFieldsMixin(object):
         return [
             c["person_or_org"]["name"] for c in obj["metadata"].get("contributors", [])
         ] or missing
-
-
-class DumperMixin(Schema):
-    """A mixin that adds post-dump and pre-dump processing functionality to the `Schema` class.
-
-    This mixin provides a way to modify the serialized or deserialized output of a schema after or before it has been dumped.
-    """
-
-    def __init__(self, dumpers=None, **kwargs):
-        """Initialize a new instance of the `DumperMixin` class.
-
-        :param dumpers: A list of `DumpProcessorMixin` instances used to modify the serialized output, defaults to None.
-        :type dumpers: list, optional
-        :raises AssertionError: If any item in the `dumpers` list is not an instance of `DumpProcessorMixin`.
-        """
-        super().__init__(**kwargs)
-        self.dumpers = dumpers or []
-        assert all(isinstance(p, FieldDumper) for p in self.dumpers)
-
-    @post_dump(pass_original=True)
-    def post_dump_pipeline(self, data, original, many, **kwargs):
-        """Applies a sequence of post-dump steps to the serialized data.
-
-        This method defines a pipeline consisting of `DumpProcessorMixin` instances that implement a `post_dump` method.
-
-        :param data: The result of serialization.
-        :param original: The original object that was serialized.
-        :param many: Whether the serialization was done on a collection of objects.
-
-        :returns: The result of the pipeline processing on the serialized data.
-        """
-        for dumper in self.dumpers:
-            # Data is assumed to be modified and returned by the dumper
-            data = dumper.post_dump(data, original)
-        return data
-
-    @pre_dump
-    def pre_dump_pipeline(self, data, many, **kwargs):
-        """Applies a sequence of pre-dump steps to the input data.
-
-        This method defines a pipeline consisting of `DumpProcessorMixin` instances that implement a `pre_dump` method.
-
-        :param data: The result of serialization.
-        :param many: Whether the serialization was done on a collection of objects.
-
-        :returns: The result of the pipeline processing on the serialized data.
-        """
-        for dumper in self.dumpers:
-            # Data is assumed to be modified and returned by the dumper
-            data = dumper.pre_dump(data)
-        return data
-
-
-class LoaderMixin(Schema):
-    def __init__(self, loaders=None, **kwargs):
-        """Initialize a new instance of the `LoaderMixin` class.
-
-        :param loaders: A list of `LoadProcessorMixin` instances used to modify the serialized output, defaults to None.
-        :type loaders: list, optional
-        :raises AssertionError: If any item in the `loaders` list is not an instance of `LoadProcessorMixin`.
-        """
-        super().__init__(**kwargs)
-        self.loaders = loaders or []
-        assert all(isinstance(p, FieldLoader) for p in self.loaders)
-
-    @post_load(pass_original=True)
-    def post_load_pipeline(self, data, original, many, **kwargs):
-        """Applies a sequence of post-load steps to the serialized data.
-
-        This method defines a pipeline consisting of `LoadProcessorMixin` instances that implement a `post_load` method.
-
-        :param data: The result of deserialization.
-        :param original: The original object that was serialized.
-        :param many: Whether the deserialization was done on a collection of objects.
-
-        :returns: The result of the pipeline processing on the deserialized data.
-        """
-        for loader in self.loaders:
-            # Data is assumed to be modified and returned by the loader
-            data = loader.post_load(data, original)
-        return data
-
-    @pre_load
-    def pre_load_pipeline(self, data, many, **kwargs):
-        """Applies a sequence of pre-dump steps to the input data.
-
-        This method defines a pipeline consisting of `DumpProcessorMixin` instances that implement a `pre_dump` method.
-
-        :param data: The result of serialization.
-        :param many: Whether the serialization was done on a collection of objects.
-
-        :returns: The result of the pipeline processing on the serialized data.
-        """
-        for dumper in self.loaders:
-            # Data is assumed to be modified and returned by the loader
-            data = dumper.pre_load(data)
-        return data
